# apps/backend/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Arenda Planform API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
...
  /properties:
    post:
      operationId: createProperty
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PropertyCreateRequest'
      responses:
        '201':
          description: Property created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/SubscriptionBlocked'
    get:
      operationId: listProperties
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Properties list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertiesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /properties/{id}:
    get:
      operationId: getProperty
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Property details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertyResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/MembershipSuspended'
    patch:
      operationId: updateProperty
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PropertyUpdateRequest'
      responses:
        '200':
          description: Property updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/SubscriptionBlocked'
    delete:
      operationId: deleteProperty
      description: >-
        Deletes the property in the given mode.
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: mode
          in: query
          required: true
          description: >-
            Deletion mode. `cascade` deletes the property outright. `detach`
            first suspends the property's billing lifecycle (same as
            archiving) and then deletes the property.
          schema:
            type: string
            enum:
              - cascade
              - detach
      responses:
        '204':
          description: Property deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/SubscriptionBlocked'
...
  /properties/{propertyId}/contacts:
    post:
      operationId: createPropertyContact
      security:
        - sessionCookie: []
      parameters:
        - name: propertyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PropertyContactCreateRequest'
      responses:
        '201':
          description: Property contact created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertyContactResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/SubscriptionBlocked'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    get:
      operationId: listPropertyContacts
      security:
        - sessionCookie: []
      parameters:
        - name: propertyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Property contacts list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertyContactsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /operations:
    get:
      operationId: listOperations
      security:
        - sessionCookie: []
      parameters:
        - name: type
          in: query
          required: false
          style: form
          explode: true
          schema:
            type: array
            items:
              $ref: '#/components/schemas/OperationType'
        - name: status
          in: query
          required: false
          style: form
          explode: true
          schema:
            type: array
            items:
              $ref: '#/components/schemas/OperationStatus'
        - name: category_id
          in: query
          required: false
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
              format: uuid
        - name: property_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: recurring_operation_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: lease_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: exclude_archived_properties
          in: query
          required: false
          schema:
            type: boolean
            default: false
        - name: sort
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/OperationListSort'
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
            maximum: 100000
      responses:
        '200':
          description: Operations list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /operations/{id}:
    get:
      operationId: getOperation
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Operation details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: updateOperation
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OperationUpdateRequest'
      responses:
        '200':
          description: Operation updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/SubscriptionBlocked'
    delete:
      operationId: deleteOperation
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Operation deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/SubscriptionBlocked'
...
  /leases:
    post:
      operationId: createLease
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeaseCreateRequest'
      responses:
        '201':
          description: Lease created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaseResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/SubscriptionBlocked'
    get:
      operationId: listLeases
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Leases list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeasesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /admin/stats:
    get:
      operationId: getAdminStats
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Platform-wide counters and recent activity for the admin dashboard
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_id
      description: >
        The actual cookie name depends on the `cookieSecure` setting:
        `session_id` when insecure and `__Host-session_id` when secure.
...
  schemas:
...
    PropertyCreateRequest:
      type: object
      required: [name, type, address]
      properties:
        name:
          type: string
          maxLength: 255
        type:
          $ref: '#/components/schemas/PropertyType'
        address:
          type: string
          maxLength: 500
        description:
          type: string
          maxLength: 2000
        attributes:
          $ref: '#/components/schemas/PropertyAttributes'
...
    PropertyContactCreateRequest:
      type: object
      required: [name, phone]
      properties:
        name:
          type: string
          maxLength: 255
        phone:
          type: string
...
    OperationType:
      type: string
      enum: [income, expense]
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    SubscriptionBlocked:
      description: Subscription is blocked; renew to continue mutating data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
...